move `setupUnlockProtocolVariable` back into lib
