+++ packages/paywall/src/index.ts
@@ -1,3 +1,4 @@
 export { Paywall } from './Paywall'
 export type { NetworkConfigs } from '@unlock-protocol/types'
 export { networkConfigs } from './networkConfigs'
+export { setupUnlockProtocolVariable } from './utils'
